pyproject.toml
[project]
name = "research-hub-pipeline"
version = "0.26.0"
description = "Zotero -> Obsidian -> NotebookLM research pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "platformdirs>=4.0",
    "pyzotero>=1.5.18",
    "pyyaml>=6",
    "rapidfuzz>=3.0",
    "requests>=2.28",
]

[project.urls]
Homepage = "https://github.com/WenyuChiou/research-hub"
Repository = "https://github.com/WenyuChiou/research-hub"
Issues = "https://github.com/WenyuChiou/research-hub/issues"

[project.optional-dependencies]
mcp = ["fastmcp>=2.0"]
playwright = ["playwright>=1.40"]
dev = [
    "pytest>=7",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "responses>=0.23",
]

[project.scripts]
research-hub = "research_hub.cli:main"
research-hub-mcp = "research_hub.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_hub"]

[tool.pytest.ini_options]
markers = [
    "stress: stress/load tests excluded from default pytest run",
    "network: Tests that hit real external APIs (OpenAlex, arXiv, etc)",
    "evals: Evaluation tests measuring search accuracy and fit-check quality",
    "slow: Tests that take >5 seconds to run",
]
testpaths = ["tests"]
addopts = "--ignore=tests/stress -p no:cacheprovider -p no:tmpdir"
